added helper object for feedbacks
