+++ app/controllers/FeedbackFormsResponseController.scala
@@ -69,8 +69,9 @@ class FeedbackFormsResponseController @Inject()(messagesApi: MessagesApi,
                     new Date(session.expirationDate.value).toString)
                 val questions = form.questions.map(questions => QuestionInformation(questions.question, questions.options))
                 val associatedFeedbackFormInformation = FeedbackForms(form.name, questions, form.active, form._id.stringify)
+                val feedbackFormQuestionOptionCount = FeedbackFormsHelper.jsonCountBuilder(form)
 
-                Some((sessionInformation, Json.toJson(associatedFeedbackFormInformation).toString))
+                Some((sessionInformation, Json.toJson(associatedFeedbackFormInformation).toString, feedbackFormQuestionOptionCount))
               case None       =>
                 Logger.info(s"No feedback form found correspond to feedback form id: ${session.feedbackFormId} for session id :${session._id}")
                 None
